requirements.txt
opencv-python
numpy
pyautogui
pillow
pystray
pygetwindow
pytesseract
requests
